requirements.txt
discord.py
python-valve
paramiko
pytz
python-dateutil
